Ct 桁サンプルで確認した C# 版の結果表示を修正しました。基本結果3画面に値を表示し、入力だけのファイルを開いたときは旧結果を消すようにしました。変位は補助節点の除外と単位換算、断面力は測点・丸め・容量上限、反力は数値精度を旧版 TS に揃えています。
同じ Ct 入力で再比較した結果、変位 COMBINE 71,064行と断面力 COMBINE 1,085,136行は生値が一致しました。変位・反力・断面力の PICKUP も、全行の表示値と選択元が一致しています。以前停止していた変位・断面力の全94 COMBINE、14 PICKUP を処理できます。

+++ FrameWebforCS/components/input/InputCombineService.cs
@@ -27,12 +27,12 @@ namespace FrameWebforCS.components.input
         // 外部からはこのプロパティを通じてのみインスタンスにアクセスできる
         public static InputCombineService Instance => _instance.Value;
 
-        private Dictionary<int, clsCombine<float>> _combine = new();
+        private Dictionary<int, clsCombine<double>> _combine = new();
         private Dictionary<int, clsCombine<int>> _define = new();
         private Dictionary<int, clsCombine<int>> _pickup = new();
 
         public IReadOnlyDictionary<int, clsCombine<int>> DefineRows => _define;
-        public IReadOnlyDictionary<int, clsCombine<float>> CombineRows => _combine;
+        public IReadOnlyDictionary<int, clsCombine<double>> CombineRows => _combine;
         public IReadOnlyDictionary<int, clsCombine<int>> PickupRows => _pickup;
         public event EventHandler? RowsReplaced;
         public event EventHandler? RowsChanged;
@@ -55,8 +55,8 @@ namespace FrameWebforCS.components.input
         /// <param name="jsonData"></param>
         public void setCombineJson(JsonElement jsonData)
         {
-            Dictionary<int, clsCombine<float>> combine =
-                JsonToDict(jsonData, "combine", ReadSingle);
+            Dictionary<int, clsCombine<double>> combine =
+                JsonToDict(jsonData, "combine", ReadDouble);
             Dictionary<int, clsCombine<int>> define =
                 JsonToDict(jsonData, "define", ReadInt32);
             Dictionary<int, clsCombine<int>> pickup =
@@ -144,11 +144,11 @@ namespace FrameWebforCS.components.input
             return combine;
         }
 
-        private static float ReadSingle(JsonElement value)
+        private static double ReadDouble(JsonElement value)
         {
             if (value.ValueKind != JsonValueKind.Number ||
-                !value.TryGetSingle(out float result) ||
-                !float.IsFinite(result))
+                !value.TryGetDouble(out double result) ||
+                !double.IsFinite(result))
             {
                 throw new JsonException("有限な数値ではありません。");
             }
@@ -173,9 +173,9 @@ namespace FrameWebforCS.components.input
                 RowsChanged?.Invoke(this, EventArgs.Empty);
         }
 
-        public void SetCombineCoefficient(int row, int column, float? value)
+        public void SetCombineCoefficient(int row, int column, double? value)
         {
-            if (value.HasValue && !float.IsFinite(value.Value))
+            if (value.HasValue && !double.IsFinite(value.Value))
                 throw new ArgumentOutOfRangeException(nameof(value));
             if (SetCoefficient(_combine, row, column, value))
                 RowsChanged?.Invoke(this, EventArgs.Empty);
